requirements.txt
# requirements.txt
flask==3.0.0
openai==1.12.0
pypdf==4.0.1
fillpdf==0.7.0
python-docx==1.1.0
openpyxl==3.1.2
chromadb==0.4.22
sentence-transformers==2.2.2
reportlab==4.0.9
fitz==0.0.1.dev2
PyMuPDF==1.23.14
python-multipart==0.0.6
Pillow==10.2.0
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.1
